create AutoGenerator script
ステージ生成に関する３つのscriptを一つにしました
->AutoGenerator.cs
