--- Assets/NakagawaTemp/TempScripts/CSVReader.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: dfd9a25b27529411c9fe62e102e64888
-timeCreated: 1455890081
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
