--- Assets/NakagawaTemp/TempScripts/Data.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: cbe65cbfe0d7c4be0b56bc6629b23730
-timeCreated: 1455890718
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
