--- Assets/NakagawaTemp/TempScripts/Generator.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 84c9194adc29e4e369eb0bcb2058e15f
-timeCreated: 1454623876
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
