+++ Assets/Resources/CSV/InputData/TestStage.csv.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4ebd99f5533d5498a92889d4e4206599
+timeCreated: 1456324863
+licenseType: Free
+TextScriptImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
