+++ Assets/pacage.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 59362f511e7b5604d85bb615ea6e6760
+folderAsset: yes
+timeCreated: 1448457983
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
